+++ Pipe.cpp
@@ -1,4 +1,4 @@
 #include "Pipe.h"
 
-Pipe::Pipe(string servicePointA, string servicePointB, unsigned long capacity, bool unidirectional)
+Pipe::Pipe(string servicePointA, string servicePointB, double capacity, bool unidirectional)
     : servicePointA(servicePointA), servicePointB(servicePointB), capacity(capacity), unidirectional(unidirectional) {}
