+++ Pipe.h
@@ -13,7 +13,7 @@ class Pipe {
 private :
     string servicePointA;
     string servicePointB;
-    unsigned long capacity;
+    double capacity;
     bool unidirectional;
 
 public:
@@ -29,7 +29,7 @@ public:
     * @param capacity The maximum capacity of the pipe.
     * @param unidirectional Boolean indicating whether the connection between source service and target service is undirected or bidirected
     */
-    Pipe(string servicePointA, string servicePointB, unsigned long capacity, bool unidirectional);
+    Pipe(string servicePointA, string servicePointB, double capacity, bool unidirectional);
 };
