+++ PumpingStation.cpp
@@ -1,3 +1,3 @@
 #include "PumpingStation.h"
 
-PumpingStation::PumpingStation(unsigned long id, string code) : id(id), code(code) {}
+PumpingStation::PumpingStation(double id, string code) : id(id), code(code) {}
